When reaping a container, supply a reason code.
We want to be able to alert users as to the reason a container
was removed. In the various conditions that trigger a reap, we
add a `reapedReason` code to the build object. This is then added
as a query parameter to the delete request that is picked up by
the container manager. Tests have been updated to accomodate the
new data structure.

Add reason for reap to delete requests.

use constants

Use query params.