Fix server tests.
The server tests were failing because we were calling real
container manager rather than mocking the response we'd expect of
it. This resulted in errors and the reap never actually
registered because the build was not thought to be deleted. Here
we add nock to mock out a 200 response from DELETE requests to
the container manager.

Also fix some syntax rearrange the requires to be in alphabetical
order.