Reaper server respects pinned builds.
This change allows the reaper server to respect (to an extent)
the notion of pinned builds. To decide what builds should be
removed, the reaper sorts builds for an organization and for
branches. The earlier items in the sorted list are more likely to
be removed. This change reorganizes the list so that pinned
builds are moved to the end of the list in date order. In this
way, the newer pinned builds are less likely to be removed than
older pinned builds. But pinned builds still live by the limits
of the organization and branch and so if all builds are pinned,
some will still be removed.

Tests are also added to ensure the expected behavior.