Reaper CLI respects pinned builds.
The criteria for reaping doesn't actually contain much of the
sorting logic that is needed to handle prioritizing pinned
builds. Much of that is handled in reaper.js which has logic to
get all builds from the CM and then loop over them, organizing
and sorting as it goes. These changes allow for the PRs that are
marked as `pinned` to be sorted so that they are less likely to
be reaped that newer builds that do not have the pinned flag.

New tests are also added to ensure that we test the reaper CLI
mode's sorting behavior which had previously not been tested.