Ensure pinned builds don't get reaped by counts.
To prevent pinned builds from causing the reaper to always
remove the newest build, we need to make sure that per branch or
per PR build limits do not take pinned builds into account. Here
we allow the limits to be set by the yaml config file and we
further remove pinned builds from data structures that are used
to determine if a build should be reaped or not. The end result
is that pinned builds will only get reaped when the disk space
limit is reached but not based on count-style limits.