Fix reaper prioritization.
The reaper's standalone mode had been improperly marking some
builds for deletion. This was a result of calculating the number
of builds on a PR before removing the pinned builds from the
list. The result is that pinned builds would never be removed but
unpinned builds that shouldn't be were being reaped. This change
removes the pinned builds from the lists before it calculates the
number of builds on the branch. Then the pinned builds are added
back to the array of items to keep.

 - Fix reap criteria logic
 - Adjust tests
 - Ensure we can set reaper criteria settings from config files