Run tests without binding to default ports.
Remove restify as it does not support dynamically selecting an
open port and it is overkill for what we're doing with http in
this service and it brings a lot of dependencies with it. When
we need more we should just use express as it is now smaller
than restify (and we were using restify as a lighter weight
alternative to express).