Adjust test.
Since we no longer use the reaper to decide how much disk space an
organization is using, we need to mock out the web requests using nock.
This is not ideal and the tests are more brittle as a result. The reason
for this is that we must know and keep track of how much disk usage
  should be reported as events are occuring. Eventually, we'll remove
  all the internal data stores from the reaper and fetch all the data
  from the db but for now we have to run things in this odd way.